requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.0
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mysql-connector-python==8.1.0
protobuf==4.21.12
python-dotenv==1.0.0
yarl==1.9.2
